lab2/django_backend/lab2_backend/requirements.txt
dj_database_url==2.1.0
Django>=3.2.1
python-decouple==3.8
django-cors-headers==4.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
